Transition inline cache and run dispatch programs
This is just enough to get the current set of tests to pass, including
those that depend on the fixation of constants implied by the dispatch
program (which thus failed up until now). It only handles callsites that
are monomorphic; if the dispatch programs fails because the site is
polymorphic, then for now we just panic about this case being NYI. There
are some missing bits still, however no tests yet trigger them.